% Generated by roxygen2 (4.0.1): do not edit by hand
\name{darken_swatch}
\alias{darken_swatch}
\title{Make ggthemr swatch darker}
\usage{
darken_swatch(amount)
}
\arguments{
\item{amount}{value between 0 and 1 indicating by how much the swatch colours should be darker/lighter.}
}
\description{
Update a ggthemr swatch so that colours are darker by some specified amount.
}
\examples{
ggthemr('dust')
darken_swatch(themr, 0.2)
}
\author{
Ciaran Tobin
}

