% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_gradient.R
\name{define_gradient}
\alias{define_gradient}
\title{Create a ggthemr_gradient object}
\usage{
define_gradient(colours, ..., low, mid, high, colors)
}
\arguments{
\item{x}{Vector of colours}
}
\value{
Object of type ggthemr_gradient
}
\description{
Given a vectors of colors, checks will be carried out to make sure colours
are not duplicated and are valid, and creates fashions values into arguments
to a known ggplot2 gradient layer See \code{\link{define_palette}} for more
details.
}
