% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_swatch.R
\name{define_swatch}
\alias{define_swatch}
\alias{ggthemr_swatch}
\title{Create a ggthemr_swatch object}
\usage{
define_swatch(x)

ggthemr_swatch(x)
}
\arguments{
\item{x}{Vector of colours}
}
\value{
Object of type ggthemr_swatch
}
\description{
Given a vectors of colors, checks will be carried out to make sure colours are not duplicated and are valid. See \code{\link{define_palette}} for more details.
}
