% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_reexports.R
\docType{data}
\name{.ggplot2}
\alias{.ggplot2}
\title{create object for ggplot2 backups}
\format{An object of class \code{list} of length 456.}
\usage{
.ggplot2
}
\description{
create object for ggplot2 backups
}
\keyword{datasets}
