% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_reexports.R
\name{.onAttach}
\alias{.onAttach}
\title{cache ggplot2 objects into ggthemr, allowing them to be modified directly
within ggplot2 and restored as needed}
\usage{
.onAttach(libname, pkgname)
}
\description{
this doesn't feel ideal, but prevents flooding of a the global namespace with
masked ggplot2 functions
}
