% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemr-package.r, R/ggthemr.R
\docType{package}
\name{ggthemr}
\alias{ggthemr}
\alias{ggthemr-package}
\title{Style ggplot2}
\usage{
ggthemr(palette = "dust", layout = "clear", spacing = 1.6,
  text_size = 12, type = "inner", line_weight = 0.5, pos = 1,
  envir = as.environment(pos))
}
\arguments{
\item{palette}{Colour scheme by name or ggtheme object.}

\item{layout}{The layout of gridlines, axes etc.}

\item{spacing}{A numeric value of how much space should exist on any given plot. Low values will result in more cramped plots, higher values will have more whitespace.}

\item{text_size}{base text size.}

\item{type}{Set to 'inner' where the background colour should not extend past strictly the plot area (or "panel" in ggplot2 terminology). 'outer' will apply the background colour to the entire plot.}

\item{line_weight}{The thickness of axes lines and gridlines. Useful for saving plots at different resolutions.}

\item{pos}{Where the new scales are stored.}

\item{envir}{The environment where the new scales are stored.}
}
\description{
ggthemr provides colour schemes and layouts for ggplot2.

Set theme to be applied to all subsequent ggplots.
}
\examples{
 ggthemr('pale', layout = 'scientific', spacing = 2, type = 'inner')
}
\author{
Ciaran Tobin
}

