% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggthemr_reset.R
\name{ggthemr_reset}
\alias{ggthemr_reset}
\title{Reset Theme}
\usage{
ggthemr_reset()
}
\description{
Remove ggthemr theme and restore the ggplot2 default.
}
\author{
Kohske Takahashi, Ciaran Tobin
}

