% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is_colour}
\alias{is_colour}
\alias{is_colour.character}
\alias{is_colour.factor}
\alias{is_colour.logical}
\alias{is_colour.numeric}
\title{Is an object a valid colour?}
\usage{
is_colour(x)

\method{is_colour}{character}(x)

\method{is_colour}{numeric}(x)

\method{is_colour}{logical}(x)

\method{is_colour}{factor}(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
Check an object to see if it is a valid colour or set of colours.
}
\details{
Colours in R can be represented by names (\code{\link{colours}}), index number (\code{\link{palette}}) or a hexidecimal character.
}
\examples{
is_colour('#ffffff')
is_colour('white')
is_colour('periwinkle blue')
is_colour(4L)
}
\author{
Ciaran Tobin
}

