% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ggthemr.R
\name{is_ggthemr}
\alias{is_ggthemr}
\title{Is a ggthemr object}
\usage{
is_ggthemr(x)
}
\arguments{
\item{x}{object to test.}
}
\description{
Test an object to determine if it is of class ggthemr.
}
\author{
Ciaran Tobin
}

