% Generated by roxygen2 (4.0.1): do not edit by hand
\name{legend_bottom}
\alias{legend_bottom}
\alias{legend_left}
\alias{legend_right}
\alias{legend_top}
\title{ggplot2 Legend Position}
\usage{
legend_bottom()

legend_top()

legend_left()

legend_right()
}
\description{
Move a ggplot2 legend to the bottom of the plot.
}
\author{
Ciaran Tobin
}

