% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lighten_palette}
\alias{lighten_palette}
\title{Make ggthemr palette lighter}
\usage{
lighten_palette(x, amount)
}
\arguments{
\item{x}{ggthemr object.}

\item{amount}{value between 0 and 1 indicating by how much the swatch colours should be darker/lighter.}
}
\description{
Update a ggthemr palette so that colours are lighter by some specified amount. This includes everything (background, axes lines, swatch colours, gradient etc.).
}
\examples{
themr <- ggthemr('pale')
lighten_palette(themr, 0.2)
}
\author{
Ciaran Tobin
}

