% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_legend.R
\name{move_legend}
\alias{move_legend}
\title{Move legend.}
\usage{
move_legend(position)
}
\arguments{
\item{position}{The location of the legend ('top', 'bottom', 'right', 'left' or 'none').}
}
\description{
move legend to given location.
}

