% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_axes.R
\name{no_axes_titles}
\alias{no_axes_titles}
\title{Remove ggplot2 axes titles}
\usage{
no_axes_titles()
}
\description{
Do not show titles for all axes
}
\seealso{
\code{\link{no_x_title}}, \code{\link{no_y_title}}
}
\author{
Ciaran Tobin
}
