% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/no_gridlines.R
\name{no_gridlines}
\alias{no_gridlines}
\title{Remove all gridlines}
\usage{
no_gridlines()
}
\description{
Remove all vertical and horizontal gridlines (major and minor) from a ggplot2 plot or theme.
}
\author{
Ciaran Tobin
}

