% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_axes.R
\name{no_ticks}
\alias{no_ticks}
\title{Remove ggplot2 tick marks}
\usage{
no_ticks()
}
\description{
Do not show tick marks for all axes
}
\seealso{
\code{\link{no_x_ticks}}, \code{\link{no_y_ticks}}
}
\author{
Ciaran Tobin
}
