% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_axes.R
\name{no_x_axis}
\alias{no_x_axis}
\title{Remove ggplot2 x-axis}
\usage{
no_x_axis()
}
\description{
Remove the x axis from a ggplot2 including the axis lines, titles, labels and ticks.
}
\author{
Ciaran Tobin
}

