% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/no_axes.R
\name{no_y_axis}
\alias{no_y_axis}
\title{Remove ggplot2 y-axis}
\usage{
no_y_axis()
}
\description{
Remove the y-axis from a ggplot2 including the axis lines, titles, labels and ticks.
}
\author{
Ciaran Tobin
}

