% Generated by roxygen2 (4.0.2): do not edit by hand
\name{no_y_gridlines}
\alias{no_horizontal_gridlines}
\alias{no_y_gridlines}
\title{Remove horizontal gridlines.}
\usage{
no_y_gridlines()

no_horizontal_gridlines()
}
\description{
Remove all major and minor horizontal gridlines from a ggplot2 plot or theme.
}
\section{Functions}{
\itemize{
\item \code{}: 
}}
\author{
Ciaran Tobin
}

