% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/no_axes.R
\name{no_y_text}
\alias{no_y_text}
\title{Remove ggplot2 y-axis text/labels}
\usage{
no_y_text()
}
\description{
Remove ggplot2 y-axis text/labels
}
\author{
Ciaran Tobin
}
\seealso{
\code{\link{no_x_text}}
}

