% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/no_axes.R
\name{no_y_title}
\alias{no_y_title}
\title{Remove ggplot2 y-axis titles}
\usage{
no_y_title()
}
\description{
Remove ggplot2 y-axis titles
}
\author{
Ciaran Tobin
}
\seealso{
\code{\link{no_axes_titles}}, \code{\link{no_x_title}}
}

