% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ggthemr}
\alias{plot.ggthemr}
\title{Create colour plot of ggthemr palette}
\usage{
\method{plot}{ggthemr_palette}(x)
}
\arguments{
\item{x}{ggthemr palette object.}
}
\description{
Create colour plot of ggthemr palette
}

