% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ggthemr_swatch}
\alias{plot.ggthemr_swatch}
\title{Create colour plot of ggthemr swatch}
\usage{
\method{plot}{ggthemr_swatch}(x, ...)
}
\arguments{
\item{x}{ggthemr swatch object.}

\item{...}{other arguments to plot() (these will be ignored).}
}
\description{
Create colour plot of ggthemr swatch
}

