% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rotate_axes_text}
\alias{rotate_axes_text}
\title{Rotate ggplot2 axes labels}
\usage{
rotate_axes_text(angle)
}
\arguments{
\item{angle}{The angle of rotation.}
}
\description{
Rotate both axes labels by a given angle.
}
\author{
Ciaran Tobin
}

