% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rotate_x_labels}
\alias{rotate_x_labels}
\title{Rotate ggplot2 x-axis labels}
\usage{
rotate_x_labels(angle, align = TRUE)
}
\arguments{
\item{angle}{The angle of rotation.}

\item{align}{}
}
\description{
Rotate the x-axis labels/text by a given angle.
}
\author{
Ciaran Tobin
}

