% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rotate_labels.R
\name{rotate_y_text}
\alias{rotate_y_text}
\title{Rotate ggplot2 x-axis labels}
\usage{
rotate_y_text(angle)
}
\arguments{
\item{angle}{The angle of rotation.}
}
\description{
Rotate the x-axis labels/text by a given angle.
}
\author{
Ciaran Tobin
}

