% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_swatch.R
\name{set_swatch}
\alias{set_swatch}
\title{Change the swatch colours in a ggthemr theme}
\usage{
set_swatch(swatch_colours)
}
\arguments{
\item{swatch_colours}{new colours to use in themr swatch}
}
\description{
Edit an existing theme's swatch. The swatch contains the main colours used in a ggplot outside of the colours for the title, axes text, axes lines, gridlines, background etc.

The first colour of the swatch is separate from the others. It is used in special cases e.g. the outline colour for geom_boxplot(). This colour should have good contrast with at least the second colour of the swatch.
}

