% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_swatch.R
\name{shuffle_swatch}
\alias{shuffle_swatch}
\title{Shuffle ggthemr swatch colours}
\usage{
shuffle_swatch()
}
\description{
Randomly reorder plot swatch colours.
}
\examples{
ggthemr('pale')
shuffle_swatch()
}
\author{
Ciaran Tobin
}

