% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_scales.R
\name{validate_gradient_args}
\alias{validate_gradient_args}
\title{Validate an argument list for required values for ggplot2::scale_*_gradient}
\usage{
validate_gradient_args(args, error = FALSE)
}
\arguments{
\item{args}{a list of arguments to provide to ggplot2::scale_*_gradient
formals}
}
\value{
\code{TRUE} if and only if \code{args} contains named values for both
  "low" and "high"
}
\description{
Validate an argument list for required values for ggplot2::scale_*_gradient
}
