% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_scales.R
\name{validate_gradientn_args}
\alias{validate_gradientn_args}
\title{Validate an argument list for required values for ggplot2::scale_*_gradientn}
\usage{
validate_gradientn_args(args, error = FALSE)
}
\arguments{
\item{args}{a list of arguments to provide to ggplot2::scale_*_gradientn
formals}
}
\value{
\code{TRUE} if and only if \code{args} contains named values for
  "colours" or "color"
}
\description{
Validate an argument list for required values for ggplot2::scale_*_gradientn
}
