% Generated by roxygen2 (4.0.2): do not edit by hand
\name{darken_gradient}
\alias{darken_gradient}
\title{Make ggthemr gradient darker}
\usage{
darken_gradient(amount)
}
\arguments{
\item{amount}{value between 0 and 1 indicating by how much the swatch colours should be darker/lighter.}
}
\description{
Update a ggthemr gradient so that colours are darker by some specified amount.
}
\examples{
ggthemr('dust')
darken_gradient(0.2)
}
\author{
Ciaran Tobin
}

