% Generated by roxygen2 (4.0.1): do not edit by hand
\name{darken_palette}
\alias{darken_palette}
\title{Make ggthemr palette darker}
\usage{
darken_palette(x, amount)
}
\description{
Update a ggthemr palette so that colours are darker by some specified amount. This includes everything (background, axes lines, swatch colours, gradient etc.).
}
\examples{
themr <- ggthemr('pale')
darken_palette(themr, 0.2)
}
\author{
Ciaran Tobin
}

