% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemr_current.R
\docType{data}
\name{ggthemr_current}
\alias{ggthemr_current}
\title{Store Theme}
\format{An object of class \code{list} of length 4.}
\usage{
ggthemr_current
}
\description{
Store ggthemr theme.
}
\keyword{datasets}
