% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/move_legend.R
\name{legend_bottom}
\alias{legend_bottom}
\alias{legend_left}
\alias{legend_right}
\alias{legend_top}
\title{ggplot2 Legend Position}
\usage{
legend_bottom()

legend_top()

legend_left()

legend_right()
}
\description{
Move a ggplot2 legend to the bottom of the plot.

Move a ggplot2 legend to the top of the plot.

Move a ggplot2 legend to the left of the plot.

Move a ggplot2 legend to the right of the plot.
}
\author{
Ciaran Tobin

Ciaran Tobin

Ciaran Tobin

Ciaran Tobin
}

