% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lighten_gradient}
\alias{lighten_gradient}
\title{Make ggthemr gradient lighter}
\usage{
lighten_gradient(amount)
}
\arguments{
\item{amount}{value between 0 and 1 indicating by how much the swatch colours should be darker/lighter.}
}
\description{
Update a ggthemr gradient so that colours are lighter by some specified amount.
}
\examples{
ggthemr('dust')
lighten_gradient(themr, 0.2)
}
\author{
Ciaran Tobin
}

