% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_contrast.R
\name{lighten_palette}
\alias{lighten_palette}
\title{Make ggthemr palette lighter}
\usage{
lighten_palette(amount)
}
\arguments{
\item{amount}{value between 0 and 1 indicating by how much the colours should be darker/lighter.}
}
\description{
Update a ggthemr palette so that colours are lighter by some specified amount. This includes everything (background, axes lines, swatch colours, gradient etc.).
}
\examples{
ggthemr('dust')
lighten_palette(0.2)
}
\author{
Ciaran Tobin
}
