% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_contrast.R
\name{lighten_swatch}
\alias{lighten_swatch}
\title{Make ggthemr swatch lighter}
\usage{
lighten_swatch(amount)
}
\arguments{
\item{amount}{value between 0 and 1 indicating by how much the colours should be darker/lighter.}
}
\description{
Update a ggthemr swatch so that colours are lighter by some specified amount.
}
\examples{
ggthemr('dust')
lighten_swatch(0.2)
}
\author{
Ciaran Tobin
}
