% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_gridlines.R
\name{no_x_gridlines}
\alias{no_vertical_gridlines}
\alias{no_x_gridlines}
\title{Remove vertical gridlines.}
\usage{
no_x_gridlines()

no_vertical_gridlines()
}
\description{
Remove all major and minor vertical gridlines from a ggplot2 plot or theme.
}
\author{
Ciaran Tobin
}

