% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_axes.R
\name{no_x_ticks}
\alias{no_x_ticks}
\title{Remove ggplot2 x-axis tick marks}
\usage{
no_x_ticks()
}
\description{
Do not show tick marks for x-axis
}
\seealso{
\code{\link{no_ticks}}, \code{\link{no_y_ticks}}
}
\author{
Ciaran Tobin
}
