% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_axes.R
\name{no_x_title}
\alias{no_x_title}
\title{Remove ggplot2 x-axis titles}
\usage{
no_x_title()
}
\description{
Do not show x-axis title
}
\author{
Ciaran Tobin
}
\seealso{
\code{\link{no_axes_titles}}, \code{\link{no_y_title}}
}

