% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/no_gridlines.R
\name{no_y_gridlines}
\alias{no_horizontal_gridlines}
\alias{no_y_gridlines}
\title{Remove horizontal gridlines.}
\usage{
no_y_gridlines()

no_horizontal_gridlines()
}
\description{
Remove all major and minor horizontal gridlines from a ggplot2 plot or theme.
}
\section{Functions}{
\itemize{
\item \code{no_horizontal_gridlines}: 
}}
\author{
Ciaran Tobin
}

