% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_axes.R
\name{no_y_ticks}
\alias{no_y_ticks}
\title{Remove ggplot2 y-axis tick marks}
\usage{
no_y_ticks()
}
\description{
Do not show tick marks for y-axis
}
\author{
Ciaran Tobin
}
\seealso{
\code{\link{no_ticks}}, \code{\link{no_x_ticks}}
}

