% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ggthemr}
\alias{plot.ggthemr}
\title{Create colour plot of ggthemr palette}
\usage{
\method{plot}{ggthemr}(x, ...)
}
\arguments{
\item{x}{ggthemr object.}

\item{...}{other arguments to plot() (these will be ignored).}
}
\description{
Create colour plot of ggthemr palette
}
