% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.ggthemr_palette}
\alias{plot.ggthemr_palette}
\title{Create colour plot of ggthemr palette}
\usage{
\method{plot}{ggthemr_palette}(x)
}
\arguments{
\item{x}{ggthemr palette object.}
}
\description{
Create colour plot of ggthemr palette
}

