% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ggthemr_palette}
\alias{plot.ggthemr_palette}
\title{Create colour plot of ggthemr palette}
\usage{
\method{plot}{ggthemr_palette}(x, ...)
}
\arguments{
\item{x}{ggthemr palette object.}

\item{...}{other arguments to plot() (these will be ignored).}
}
\description{
Create colour plot of ggthemr palette
}
