% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reverse_gradient}
\alias{reverse_gradient}
\title{Reverse ggthemr gradient colours}
\usage{
reverse_gradient()
}
\description{
Flip the colours in a ggthemr gradient.
}
\examples{
ggthemr()
reverse_gradient(themr)
}
\author{
Ciaran Tobin
}

