% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_gradient.R
\name{reverse_gradient}
\alias{reverse_gradient}
\title{Reverse ggthemr gradient colours}
\usage{
reverse_gradient()
}
\description{
Flip the colours in a ggthemr gradient.
}
\examples{
ggthemr()
reverse_gradient()
}
\author{
Ciaran Tobin
}
