% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_labels.R
\name{rotate_axes_text}
\alias{rotate_axes_text}
\title{Rotate ggplot2 axes labels}
\usage{
rotate_axes_text(angle)
}
\arguments{
\item{angle}{The angle of rotation.}
}
\description{
Rotate both axes labels by a given angle.
}
\author{
Ciaran Tobin
}

