% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_ggthemr_d.R
\name{scale_colour_ggthemr_d}
\alias{scale_colour_ggthemr_d}
\title{ggthemr colour scale}
\usage{
scale_colour_ggthemr_d(...)
}
\description{
A utility function for applying the currently selected ggthemr palette to a ggplot object
}
\examples{
ggplot(mtcars, aes(mpg, disp, colour = factor(am))) + geom_point() + scale_colour_ggthemr_d()

}
