% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_swatch.R
\name{set_swatch}
\alias{set_swatch}
\title{Change the swatch colours in a ggthemr theme}
\usage{
set_swatch(swatch_colours)
}
\arguments{
\item{swatch_colours}{new colours to use in themr swatch}
}
\description{
Edit an existing theme's swatch. The swatch contains the main colours used in a ggplot outside of the colours for the title, axes text, axes lines, gridlines, background etc.
}

