% Generated by roxygen2 (4.0.1): do not edit by hand
\name{shuffle_swatch}
\alias{shuffle_swatch}
\title{Shuffle ggthemr swatch colours}
\usage{
shuffle_swatch(x)
}
\arguments{
\item{x}{ggthemr object.}
}
\description{
Randomly reorder plot swatch colours.
}
\examples{
themr <- ggthemr('pale')
shuffle_swatch(themr)
}
\author{
Ciaran Tobin
}

