% Generated by roxygen2 (4.0.1): do not edit by hand
\name{swatch}
\alias{swatch}
\alias{swatch.ggthemr}
\alias{swatch.ggthemr_palette}
\title{Retrieve the swatch colours from a ggthemr object}
\usage{
swatch(x)

\method{swatch}{ggthemr}(x)

\method{swatch}{ggthemr_palette}(x)
}
\arguments{
\item{x}{ggthemr or ggthemr palette object.}
}
\description{
The swatch is the main set of colours used for plotting data in a ggplot2 plot. This definition excludes the colours used for the plot background, title/axis text, gridlines etc.
}
\examples{
themr <- ggthemr('sea')
swatch(themr)
}

