% Generated by roxygen2 (4.0.1): do not edit by hand
\name{swatch}
\alias{swatch}
\title{Retrieve the swatch colours from a ggthemr object}
\usage{
swatch()
}
\arguments{
\item{x}{ggthemr object.}
}
\description{
The swatch is the main set of colours used for plotting data in a ggplot2 plot. This definition excludes the colours used for the plot background, title/axis text, gridlines etc.
}
\examples{
themr <- ggthemr('dust')
swatch(themr)
}

