% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swatch.R
\name{swatch}
\alias{swatch}
\title{Retrieve the swatch colours from a ggthemr object}
\usage{
swatch()
}
\description{
The swatch is the main set of colours used for plotting data in a ggplot2 plot. This definition excludes the colours used for the plot background, title/axis text, gridlines etc.
}
\examples{
ggthemr('dust')
}

